feat : 레벨업 모션 컴포넌트 분리 #158
