+++ src/domains/MyPage/types/mission.ts
@@ -6,12 +6,3 @@ export interface MissionType {
   requireValue: number;
   expReward: number;
 }
-
-export interface ExpResultType {
-  exp: number;
-  level: number;
-  levelUpdated: boolean;
-  prevExp: number;
-  expReward: number;
-  missionName: string;
-}
