+++ src/types/expResult.ts
@@ -1,0 +1,8 @@
+export interface ExpResultType {
+  exp: number;
+  level: number;
+  levelUpdated: boolean;
+  prevExp: number;
+  expReward: number;
+  missionName: string;
+}
